# docs/openapi.yml
openapi: 3.0.0
info:
  title: "Todo list application"
  version: "1.0.0"
...
paths:
...
  /api/admin/users/{id}:
      get:
        tags:
          - Admin
        summary: Get User By ID
        description: Retrieves a user profile by query parameter ID. Verifies OTP before retrieving.
        security:
          - BearerAuth: [ ]
        parameters:
          - name: user_id
            in: query
            required: true
            schema:
              type: string
        responses:
          '200':
            description: User retrieved
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ActionResultModel'
      delete:
        tags:
          - Admin
        summary: Delete User By ID
        description: Deletes a user by path parameter user_id.
        security:
          - BearerAuth: [ ]
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string
        responses:
          '200':
            description: User deleted
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ActionResult'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication requires a JWT token which is parsed into "claims" containing the "user_id".
  schemas:
...
    ActionResult:
        type: object
        properties:
          StatusCode:
            type: integer
          IsSuccess:
            type: boolean
          Error:
            type: array
            items:
              type: string
    ActionResultModel:
        type: object
        properties:
          Data:
            type: object
          StatusCode:
            type: integer
          IsSuccess:
            type: boolean
          Error:
            type: array
            items:
              type: string
...